+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Deploying an external file with a non-whitelisted path now fails with `FAIL_ACC_DENIED_EXT_FILE` instead of
+  `FAIL_UNKNOWN_ERROR`
+
 ## [1.35.0-rc.1] - 2026-08-13
 
 ### Added